Refactor date and util method in their own modules
