+++ lib/util.js
@@ -1,0 +1,19 @@
+var _ = require('lodash');
+
+exports.dashed = dashed;
+exports.filterFloat = filterFloat;
+
+function dashed(val) {
+    var splitted = val.split('-');
+    return _.map(splitted, function(elem) {
+        return elem.trim();
+    });
+}
+
+function filterFloat(value) {
+    if (/^(\-|\+)?([0-9]+(\.[0-9]+)?|Infinity)$/.test(value)) {
+      return Number(value);
+    }
+    return NaN;
+}
+
